requirements.txt
streamlit
folium
streamlit_folium
plotly.graph_objects    
pandas
geopandas
matplotlib.pyplot